Further refactored to support multiple dataset servers
Added PUDL dataset to exercise faceted browsing, clustermaps, and other features of Dataset.  Because of the size of pudl.sqlite, it's not included here, but can be grabbed from catalyst.coop/pudl with a small amount of effort.

Also further factored/refactored Dockerfiles to support multiple datasette servers (each on their own port).  Updated notebook to show direct simultaneous connections to such different servers.  (Note, it is a recent feature of Datasette to be able to simultaneously manage up to 10 different DB files within a single instance.  We are demonstrating multiple, single-database instances.)